feat(examples): update web-multi-store to align with the latest RFC
